requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
python-dotenv==1.0.1
sqlparse==0.5.3
